+++ database/migrations/2023_10_17_232801_create_qr_codes_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('qr_codes', function (Blueprint $table) {
             $table->id();
-            $table->integer('document');
+            $table->bigInteger('document');
             $table->boolean('is_user')->default(false);
             $table->text('code_qr');
             $table->string('code');
